refactor: 패키지 네이밍 manager로 변경
